+++ lang/en/auth_ldap_syncplus.php
@@ -54,4 +54,5 @@ $string['userentriestosuspend'] = 'User entries to be suspended: {$a}';
 $string['userentriestosuspenddone'] = 'Transaction complete – User entries suspended: {$a}';
 $string['userentriestoupdate'] = 'User entries to be checked for updates: {$a}';
 $string['userentriestoupdatedone'] = 'Transaction complete – User entries updated: {$a->updated}, user entries skipped: {$a->skipped}';
+$string['usernotfoundanymore'] = 'The temporary user {$a->name} was not found in LDAP anymore, it will be skipped.';
 $string['waitinginremovalqueue'] = 'Waiting in removal queue for {$a->days} day grace period: {$a->name} ID {$a->id}';